add cli arg type helpers module
